feat (home): added texts and buttons
Refines the homepage with a new introductory paragraph, providing a clearer overview of the company's services and philosophy.

Expands the descriptions for CNC and klempířská výroba cards to better communicate the services offered.

Makes the "Kontaktujte nás!" card a link to the contact section for improved user navigation.

Adds call-to-action buttons for service requests and company information to improve user engagement.